reorganize files within the repository
