+++ functions/calculator.js
