+++ functions/capitalize.js
