+++ tests/capitalize.test.js
@@ -1,4 +1,4 @@
-const capitalize = require('./capitalize');
+const capitalize = require('../functions/capitalize');
 
 it('capitalizes the first letter', () => {
   expect(capitalize('cat')).toBe('Cat');
