+++ tests/reverse.test.js
@@ -1,4 +1,4 @@
-const reverseString = require('./reverse');
+const reverseString = require('../functions/reverse');
 
 it('reverse the string sentence', () => {
   expect(reverseString('strawberry')).toBe('yrrebwarts');
